[DSCP-122] Rearrange modules, extract Real part
