+++ src/Dscp/Educator/Secret/Real.hs
@@ -1,0 +1,1 @@
+{-# OPTIONS_GHC -F -pgmF autoexporter #-}
