+++ src/Dscp/Educator/Secret/Real/Functions.hs
@@ -1,6 +1,6 @@
 -- | Functions to work with educator secret key storage.
 
-module Dscp.Educator.Secret.Store
+module Dscp.Educator.Secret.Real.Functions
     ( linkStore
     ) where
 
@@ -9,14 +9,32 @@ import qualified Data.ByteString.Lazy as LBS
 import System.Directory (doesFileExist)
 import System.FilePath ((</>))
 
-import Dscp.Crypto (PassPhrase, keyGen, runSecureRandom)
-import Dscp.Educator.Secret.Error (EducatorSecretError (..), rewrapSecretIOError)
-import Dscp.Educator.Secret.Functions (fromEducatorSecretJson, toEducatorSecretJson)
-import Dscp.Educator.Secret.Types (EducatorSecret (..), EducatorSecretJson,
-                                   EducatorSecretParams (..))
+import Dscp.Crypto (PassPhrase, decrypt, encrypt, keyGen, runSecureRandom)
+import Dscp.Educator.Secret.Real.Error (EducatorSecretError (..), rewrapSecretIOError)
+import Dscp.Educator.Secret.Real.Types (EducatorSecret (..), EducatorSecretJson (..),
+                                        EducatorSecretParams (..), KeyfileContent)
 import Dscp.Resource.AppDir (AppDirectory (..))
 import Dscp.Util (leftToThrow)
 import Dscp.Util.Aeson (Versioned (..))
+
+---------------------------------------------------------------------
+-- Conversions
+---------------------------------------------------------------------
+
+toEducatorSecretJson :: PassPhrase -> EducatorSecret -> EducatorSecretJson
+toEducatorSecretJson pp EducatorSecret{..} =
+    let esjEncSecretKey = encrypt pp esSecretKey
+    in EducatorSecretJson{..}
+
+fromEducatorSecretJson :: MonadThrow m => PassPhrase -> EducatorSecretJson -> m EducatorSecret
+fromEducatorSecretJson pp EducatorSecretJson{..} = do
+    esSecretKey <- decrypt pp esjEncSecretKey
+        & leftToThrow SecretWrongPassPhraseError
+    return EducatorSecret{..}
+
+---------------------------------------------------------------------
+-- Storage operations
+---------------------------------------------------------------------
 
 -- | Where keyfile would lie.
 storePath :: EducatorSecretParams -> AppDirectory -> FilePath
@@ -35,16 +53,18 @@ genStore = runSecureRandom $ do
 readStore :: (MonadIO m, MonadCatch m) => FilePath -> PassPhrase -> m EducatorSecret
 readStore path pp = do
     content <- rewrapSecretIOError $ LBS.readFile path
-    Versioned mid <- eitherDecode' @(Versioned EducatorSecretJson) content
+    Versioned mid <- eitherDecode' @KeyfileContent content
         & leftToThrow (SecretDeserialisationError . toText)
     fromEducatorSecretJson pp mid
 
 -- | Write given content to store.
-writeStore :: (MonadIO m, MonadCatch m) => FilePath -> PassPhrase -> EducatorSecret -> m ()
+writeStore
+    :: (MonadIO m, MonadCatch m)
+    => FilePath -> PassPhrase -> EducatorSecret -> m ()
 writeStore path pp store =
     rewrapSecretIOError $
     LBS.writeFile path $
-    encode @(Versioned EducatorSecretJson) $
+    encode @KeyfileContent $
     Versioned $ toEducatorSecretJson pp store
 
 -- | Creates new store under given path.
@@ -62,6 +82,8 @@ createStore path pp = do
             return store
 
 -- | Syncs with store. For now store is read-only, thus it's just read.
+-- Store is also created (and assumed to be absent before this function call) if
+-- dedicated flag is passed.
 linkStore
     :: (MonadIO m, MonadCatch m)
     => EducatorSecretParams -> AppDirectory -> m EducatorSecret
